+++ src/scraper/examples/remote.py
@@ -9,7 +9,7 @@ from typing import Self
 
 import websockets
 
-from dataHandle import Agent, D_Agent
+from data_handle import Agent, D_Agent
 from src.scraper.scraper import WebsiteScraper
 from src.scraper.scraper_error import CreateButRequestFail, CreateByLackAgent
